pdf_converterr.py
# Copied from original workspace: PDF converter and pipeline
import re
import pandas as pd
from fpdf import FPDF
import os

# (File content identical to the current project file.)

# For convenience, import the implementation from the Sampling workspace
# so that the tests exercise the same code. If you want this file to be
# independent, copy the implementation directly here.
from pathlib import Path
source = Path(r"c:/Users/roger.brook/Python/Sampling/pdf_converterr.py")
if source.exists():
    # load the file content so this module contains the same code for tests
    with open(source, 'r', encoding='utf-8') as f:
        exec(f.read(), globals())
else:
    raise FileNotFoundError('Original pdf_converterr.py not found at expected location')
