requirements.txt
pandas>=1.0.0
openpyxl>=3.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
fpdf2>=2.7.0
pytest>=7.0.0
